+++ src/data/v2/mod.rs
@@ -1,11 +1,13 @@
-// Copyright (C) 2021 The apca Developers
+// Copyright (C) 2021-2022 The apca Developers
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 mod unfold;
 
 /// Definitions for retrieval of market data bars.
 pub mod bars;
-/// Functionality for retrieval of most recent quotes.
+/// Functionality for retrieval of the most recent quote.
 pub mod last_quote;
+/// Functionality for retrieving historic quotes.
+pub mod quotes;
 /// Definitions for real-time streaming of market data.
 pub mod stream;