+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Added `api::v2::calendar` module for retrieving historic and future
   market open and close times
+- Added support for historic quote retrieval via `data::v2::quotes`
 
 
 0.22.2